stackoverflow_answers/python/how can i parse free-text time intervals/pyproject.toml
[tool.poetry]
name = "answer"
version = "0.1.0"
description = ""
authors = ["Matthew Willcockson <matthew@willcockson.family>"]

[tool.poetry.dependencies]
python = "^3.8"
python-dateutil = "^2.8.1"
lark-parser = "^0.9.0"
parsedatetime = "^2.6"

[tool.poetry.dev-dependencies]
mypy = "^0.782"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
